pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "A machine learning project to predict european energy prices"
authors = [
  { name = "smnfrs" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "catboost>=1.2",
    "fredapi>=0.5",
    "holidays>=0.40",
    "ipython",
    "jupyterlab",
    "lightgbm>=4.5",
    "loguru>=0.7",
    "matplotlib>=3.9",
    "mlflow>=3.0",
    "notebook",
    "numpy>=2.1",
    "optuna>=4.0",
    "pandas>=2.2",
    "prophet>=1.1",
    "python-dotenv>=1.0",
    "ruff>=0.9",
    "scikit-learn>=1.5",
    "statsmodels>=0.14",
    "tqdm>=4.66",
    "typer>=0.15",
    "xgboost>=3.0",
    "yfinance>=0.2",
]
requires-python = "==3.13.5"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
